Integrate VS2015 support into 4.10 branch. Clean up issues as a result of gaps in required support from 4.9 (vs main).
#jira UEPLAT-1067

CLs Integrated:
2613833
2613835
2618837
2618838
2618840
2618844
2622232
2622282
2622291
2622293
2622294
2622305
2624515
2624516
2624517
2642715
2647025
2647026
2647801
2647802
2647803
2647804
2647812
2647826
2647831
2647837
2647838
2647853
2648172
2648175
2648210
2651883
2651893
2651897
2652386
2653316
2653635
2653680
2657131
2666962
2668208
2668311
2668340
2668897
2688492
2688751
2691248
2692028
2707678
2709701
2709859
2710400

[CL 2716009 by Jeff Campeau in 4.10 branch]

+++ Engine/Source/Runtime/Windows/D3D12RHI/Private/D3D12StateCache.cpp
@@ -628,7 +628,7 @@ void FD3D12DefaultBufferPool::ReleaseAllResources()
 // Grab a buffer from the available buffers or create a new buffer if none are available
 HRESULT FD3D12DefaultBufferAllocator::AllocDefaultResource(const D3D12_RESOURCE_DESC& Desc, D3D12_SUBRESOURCE_DATA *pInitialData, FD3D12ResourceLocation *ResourceLocation, uint32 Alignment)
 {
-	check (Desc.Flags < MAX_DEFAULT_POOLS);
+	check ((uint32)Desc.Flags < MAX_DEFAULT_POOLS);
 
 	return DefaultBufferPools[Desc.Flags]->AllocDefaultResource(Desc, pInitialData, ResourceLocation, Alignment);
 }
